cart-service/requirements.txt
Django==5.1.5
djangorestframework==3.15.2
PyJWT==2.9.0
PyMySQL==1.1.1
djangorestframework-simplejwt==5.4.0
